Default app can be one that wasn't specified
Previously, not having the default app for a given handler installed
resulted in the default being the fallback, even if another app was
available. This causes the default app to fall back to the first
available app if not the default.

(For now, this is really only relevant to Twitter, the only handler both
with no first-party default and with multiple handlers)