WIP: Support for CityMapper
There are two issues stopping me from merging this in:

1. The example project is currently broken, so I don't have good way to
verify functionality. This is a deep rabbithole I'll hopefully have time
to go down soon.

2. The CityMapper URL scheme only lets you link to directions with
lat/lng coordinates. Currently, an explicit API contract of the
INKMapsHandler is that plaintext street address / point of interest
strings are okay. I'm not sure I want to merge this in as long as that
tension exists.